refactor: move recurrence code into common module
